+++ rouver_test/handler.py
@@ -1,8 +1,9 @@
 from http import HTTPStatus
-from typing import Type, Iterator
+from typing import Type, Iterable
 from unittest import TestCase
 
 from asserts import assert_is, assert_equal
+
 from werkzeug.wrappers import Request
 
 from rouver.handler import RouteHandlerBase
@@ -21,15 +22,19 @@ class RouteHandlerBaseTest(TestCase):
         return b"".join(handler)
 
     def test_attributes(self) -> None:
-        handler = RouteHandlerBase(self.request, ["foo"], self.start_response)
+        class TestingHandler(RouteHandlerBase):
+            def prepare_response(self) -> Iterable[bytes]:
+                pass
+        handler = TestingHandler(self.request, ["foo"], self.start_response)
         assert_is(self.request, handler.request)
         assert_equal(["foo"], handler.path_args)
         assert_is(self.start_response, handler.start_response)
 
     def test_respond(self) -> None:
         class TestingHandler(RouteHandlerBase):
-            def __iter__(self) -> Iterator[bytes]:
-                return self.respond()
+            def prepare_response(self) -> Iterable[bytes]:
+                self.respond()
+                return [b"foo", b"bar"]
         response = self.call_handler(TestingHandler)
         self.start_response.assert_status(HTTPStatus.OK)
-        assert_equal(b"", response)
+        assert_equal(b"foobar", response)